🎈 feat(mobile/entities/theme): reorganize theme management by moving to entities and creating new theme atom
