+++ apps/mobile/src/layers/app/providers/auth-reset-provider.tsx
@@ -3,9 +3,9 @@ import { type WritableAtom, useStore } from "jotai";
 import { RESET } from "jotai/utils";
 import { type FC, type ReactNode, useEffect } from "react";
 import { accountAssociationSessionTokenAtom, sessionTokenAtom } from "../../entities/session";
+import { themeAtom } from "../../entities/theme";
 import { userAtom } from "../../entities/user";
 import { visitPersonalizePageFlagAtom } from "../../features/auth";
-import { themeAtom } from "../../features/theme";
 import { subscribeToAuthReset } from "../../shared/lib/auth";
 import type { globalStorageKeys, secureStorageKeys } from "../../shared/lib/storage";
 import { useResetJotaiStore } from "./resettable-jotai-provider";
