+++ apps/mobile/src/layers/entities/theme/index.ts
@@ -1,0 +1,4 @@
+export { useTheme } from "./model/use-theme";
+export { themeAtom } from "./model/theme-atom";
+
+export type { Theme } from "./model/theme-atom";
