+++ apps/mobile/src/layers/entities/theme/model/theme-atom.ts
@@ -1,4 +1,4 @@
-import { atomWithGlobalStorage, globalStorageKeys } from "../../../shared/lib";
+import { atomWithGlobalStorage, globalStorageKeys } from "../../../shared/lib/storage";
 
 export type Theme = "light" | "dark" | "system";
