--- apps/mobile/src/layers/features/theme/index.ts
@@ -1,4 +1,0 @@
-export { useTheme } from "./lib/use-theme";
-export { themeAtom } from "./model/theme";
-
-export type { Theme } from "./model/theme";
